Move Ethereum Payments classes to golem.ethereum.
